requirements-dev.txt
# Development dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-playwright>=0.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
pre-commit>=3.0.0

# Testing
pytest-cov>=4.0.0

# Documentation
mkdocs>=1.5.0
mkdocs-material>=9.0.0